began SelConn and BnetGW

+++ DiabloUI/disclaim.cpp
@@ -21,7 +21,7 @@ LRESULT __stdcall disclaim_WndProc(HWND hWnd, UINT Msg, WPARAM wParam, LPARAM lP
 					disclaim_FadeFromDisclaim(hWnd);
 			} else if (Msg == 2024) {
 				if (!Fade_CheckRange5())
-					Fade_SetFadeTimer((int)hWnd);
+					Fade_SetFadeTimer( hWnd );
 				return 0;
 			}
 			return (LRESULT)SDlgDefDialogProc(hWnd, Msg, (HDC)wParam, (HWND)lParam);
